Link protobuf statically (#14)
* build protobuf statically into the bindings; still needs an update to the manylinux image

* restore previous patch since it seems valhalla doesn't need to statically link to protobuf, only the bindings valhalla library has to

* change pre-commit black version to be compatible with latest click version

* 'extra_objects' belong in the extension of course

* update valhalla and protobuf

* check in static protobuf library for linux

* updated valhalla on windows; couldn't make a static lib linking work for protobuf, maybe it's because libvalhalla is linked dynamically; in any case: seems like QGIS 3.24 is happy with the dynamic protobuf link as well, no weird errors

* update mac

+++ include/darwin/geos/operation/overlay/OverlayOp.h
@@ -264,7 +264,7 @@ private:
      * but in the original arg Geometry it is actually
      * in the interior due to the Boundary Determination Rule)
      */
-    void copyPoints(uint8_t argIndex, const geom::Envelope* env = nullptr);
+    void copyPoints(int argIndex, const geom::Envelope* env = nullptr);
 
     /** \brief
      * Compute initial labelling for all DirectedEdges at each node.
@@ -309,7 +309,7 @@ private:
     /** \brief
      * Label an isolated node with its relationship to the target geometry.
      */
-    void labelIncompleteNode(geomgraph::Node* n, uint8_t targetIndex);
+    void labelIncompleteNode(geomgraph::Node* n, int targetIndex);
 
     /** \brief
      * Find all edges whose label indicates that they are in the result
@@ -389,7 +389,7 @@ private:
     double avgz[2];
     bool avgzcomputed[2];
 
-    double getAverageZ(uint8_t targetIndex);
+    double getAverageZ(int targetIndex);
     static double getAverageZ(const geom::Polygon* poly);
 
     ElevationMatrix* elevationMatrix;
